kmail, kontact: wrap external executable dependencies
follow-up to #108824

It turns out that the wrapping done for each KDE PIM component needs to
be done for the component itself as well as for kontact, which then
aggregates all of these components as KParts.
This commit introuduces specifying these wrapping dependencies in a
separate file to make them available for import from their respective
derivations.
It is designed to be extendable in case other PIM components turn out to
require external executables as well.